feat(auth): 实现管理后台登录功能
- 新增 authenticate 方法用于验证用户名和密码
- 更新 login 方法，使用新的 authenticate 方法进行身份验证
- 添加 logger 模块，实现登录日志和操作日志服务
- 优化 user模块，增加根据用户名获取用户信息的方法
- 更新 auth 模块，引入新实现的 logger 和 user 模块

+++ crates/daoyi-entities-system/src/entity/system_users.rs
@@ -19,7 +19,7 @@ pub struct Model {
     pub mobile: Option<String>,
     pub sex: Option<i16>,
     pub avatar: Option<String>,
-    pub status: i16,
+    pub status: i32,
     pub login_ip: Option<String>,
     pub login_date: Option<DateTime>,
     pub creator: Option<String>,
